+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Unreleased
   Please use the :meth:`dispersion` method of the appropriate family instance instead.
 - All functions now use ``sample_weight`` as a keyword instead of ``weights``, in line with scikit-learn.
 - All functions now use ``dispersion`` as a keyword instead of ``phi``.
+- The :meth:`tear_down_from_fit` method of :class:`~quantcore.glm.GeneralizedLinearRegressor` and :class:`~quantcore.glm.GeneralizedLinearRegressorCV` has been taken private.
 
 **New features:**